backend/requirements.txt
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
djangorestframework-jsonapi
djangorestframework-simplejwt
django-extensions==3.2.1
django-filter==22.1
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.12
gunicorn
psycopg2-binary
inflection==0.5.1
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.6
pillow